[02] create account - create code viewModel
