--- petJournal/petJournal/Features/UnAuthentication/CreateAccount/CreateAccountVM.swift
@@ -1,8 +1,0 @@
-//
-//  CreateAccountVM.swift
-//  petJournal
-//
-//  Created by Marcylene Barreto on 24/03/23.
-//
-
-import Foundation
